Move domain::buildbot to domain::builds
