+++ src/domain/mod.rs
@@ -1,5 +1,5 @@
 // Copyright 2016 Adam Perry. Dual-licensed MIT and Apache 2.0 (see LICENSE files for details).
-pub mod buildbot;
+pub mod builds;
 pub mod github;
 pub mod releases;
 pub mod rfcbot;
